+++ src/main/java/raven/datetime/component/time/AnimationChange.java
@@ -25,14 +25,19 @@ public class AnimationChange {
         animator.setInterpolator(CubicBezierEasing.EASE);
     }
 
-    public void start(float angleTarget, float marginTarget) {
+    public void start(float angleTarget, float marginTarget, boolean animated) {
         if (angle != angleTarget || margin != marginTarget) {
             angleValue.set(angle, angleTarget);
             marginValue.set(margin, marginTarget);
             if (animator.isRunning()) {
                 animator.stop();
             }
-            animator.start();
+            if (animated) {
+                animator.start();
+            } else {
+                angle = angleValue.interpolate(1f);
+                margin = marginValue.interpolate(1f);
+            }
         }
     }